requirements.txt
Flask==3.0.0
python-telegram-bot==20.7
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0